refactor(shipment): 계층별 DTO 종류별 패키지 분리
- presentation layer: request, response
- application layer: command, result
