+++ shipment-service/src/main/java/com/shipflow/shipmentservice/application/ShipmentService.java
@@ -8,13 +8,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.shipflow.common.exception.BusinessException;
-import com.shipflow.shipmentservice.application.dto.ShipmentResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteUpdateCommand;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteUpdateResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentSearchResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentUpdateCommand;
-import com.shipflow.shipmentservice.application.dto.ShipmentUpdateResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentRouteResult;
+import com.shipflow.shipmentservice.application.dto.command.ShipmentRouteUpdateCommand;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentRouteUpdateResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentSearchResult;
+import com.shipflow.shipmentservice.application.dto.command.ShipmentUpdateCommand;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentUpdateResult;
 import com.shipflow.shipmentservice.domain.Shipment;
 import com.shipflow.shipmentservice.domain.ShipmentRoute;
 import com.shipflow.shipmentservice.domain.exception.ShipmentErrorCode;
