+++ shipment-service/src/main/java/com/shipflow/shipmentservice/application/dto/result/ShipmentSearchResult.java
@@ -1,4 +1,4 @@
-package com.shipflow.shipmentservice.application.dto;
+package com.shipflow.shipmentservice.application.dto.result;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
