+++ shipment-service/src/main/java/com/shipflow/shipmentservice/presentation/dto/response/ShipmentSearchResDto.java
@@ -1,9 +1,9 @@
-package com.shipflow.shipmentservice.presentation.dto;
+package com.shipflow.shipmentservice.presentation.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.UUID;
 
-import com.shipflow.shipmentservice.application.dto.ShipmentSearchResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentSearchResult;
 import com.shipflow.shipmentservice.domain.ShipmentStatus;
 
 import lombok.Builder;
