+++ shipment-service/src/test/java/com/shipflow/shipmentservice/application/ShipmentServiceTest.java
@@ -18,13 +18,13 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 
 import com.shipflow.common.exception.BusinessException;
-import com.shipflow.shipmentservice.application.dto.ShipmentResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteUpdateCommand;
-import com.shipflow.shipmentservice.application.dto.ShipmentRouteUpdateResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentSearchResult;
-import com.shipflow.shipmentservice.application.dto.ShipmentUpdateCommand;
-import com.shipflow.shipmentservice.application.dto.ShipmentUpdateResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentRouteResult;
+import com.shipflow.shipmentservice.application.dto.command.ShipmentRouteUpdateCommand;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentRouteUpdateResult;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentSearchResult;
+import com.shipflow.shipmentservice.application.dto.command.ShipmentUpdateCommand;
+import com.shipflow.shipmentservice.application.dto.result.ShipmentUpdateResult;
 import com.shipflow.shipmentservice.domain.Shipment;
 import com.shipflow.shipmentservice.domain.ShipmentRoute;
 import com.shipflow.shipmentservice.domain.ShipmentRouteStatus;
